0.2.6.1 bugfixes, fine tuning

+++ src/main/kotlin/Action.kt
@@ -2,6 +2,6 @@ import java.util.*
 
 val actions = LinkedList<Action>()
 interface Action {
-  fun conditions(x: Int, y: Int): Boolean = true
-  fun execute(x: Int, y: Int)
+  fun conditions(x: Double, y: Double): Boolean = true
+  fun execute(x: Double, y: Double)
 }
