+++ src/main/kotlin/mod/actions/deleteShapes.kt
@@ -5,7 +5,7 @@ import mod.dragging.selectedShapes
 import mod.dragging.shapes
 
 object deleteShapes: Action {
-  override fun execute(x: Int, y: Int) {
+  override fun execute(x: Double, y: Double) {
     shapes.removeAll(selectedShapes)
     selectedShapes.clear()
   }
